Update Tower Animation (NOT WORKING since the place tower logic might need change as well)
I have updated the tower prefabs of TowerType2, TowerType3, and TowerType4 with new animations and fire points. But there is no animation in during the game play since our logic of place tower and select tower may need to change as well. But this will not break down our current game.